pyproject.toml
[project]
name = "ocr-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdown>=5.2.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
